backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
groq>=0.11.0
httpx==0.27.2
numpy==1.24.3
pandas==2.0.3
pypdf==4.0.1
duckduckgo-search==5.3.0
tavily-python==0.3.0
email-validator==2.1.0
psycopg2-binary==2.9.9